Add feature to let users download fasta files from GENCODE.
fasta.Maker has also been added. It's a generally useful class, but it 
hasn't undergone the rewrite or testing that the other classes have. It 
also isn't directly used in the main pipeline.

One downside to these added tests is that they're slow. I don't know 
that enough future development will occur on this package for it to 
matter, but avoiding pytests with downloads seems like a good dev. 
option.